pipeline/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
beautifulsoup4==4.14.2
bertopic==0.17.3
blis==1.3.3
catalogue==2.0.10
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
confection==0.1.5
cymem==2.0.13
dnspython==2.8.0
emoji==2.15.0
fasttext-wheel==0.9.2
filelock==3.20.0
fsspec==2025.10.0
h11==0.16.0
hdbscan==0.8.40
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
llvmlite==0.45.1
MarkupSafe==3.0.3
mpmath==1.3.0
murmurhash==1.0.15
narwhals==2.12.0
networkx==3.6
numba==0.62.1
numpy==1.26.4
ollama==0.6.1
packaging==25.0
pandas==2.3.3
pillow==12.0.0
plotly==6.5.0
preshed==3.0.12
pybind11==3.0.1
pydantic==2.12.5
pydantic_core==2.41.5
pymongo==4.15.4
pynndescent==0.5.13
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.16.3
sentence-transformers==5.1.2
six==1.17.0
smart_open==7.5.0
sniffio==1.3.1
soupsieve==2.8
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
sympy==1.14.0
thinc==8.3.10
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
torch==2.9.1
tqdm==4.67.1
transformers==4.57.3
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
umap-learn==0.5.9.post2
urllib3==2.5.0
wasabi==1.1.3
weasel==0.4.3
wrapt==2.0.1
